requirements.txt
streamlit==0.64.0
tensorflow==2.0.0
numpy==1.19.1
pillow==6.2.1
